fixes for 32-bit Linux

+++ example/src/libc_wrappers.rs
@@ -29,7 +29,7 @@ pub fn opendir(path: OsString) -> Result<u64, libc::c_int> {
 }
 
 pub fn readdir(fh: u64) -> Result<Option<libc::dirent>, libc::c_int> {
-    let dir: *mut libc::DIR = unsafe { mem::transmute(fh) };
+    let dir: *mut libc::DIR = unsafe { mem::transmute(fh as usize) };
     let mut entry: libc::dirent = unsafe { mem::zeroed() };
     let mut result: *mut libc::dirent = ptr::null_mut();
 
@@ -46,7 +46,7 @@ pub fn readdir(fh: u64) -> Result<Option<libc::dirent>, libc::c_int> {
 }
 
 pub fn closedir(fh: u64) -> Result<(), libc::c_int> {
-    let dir: *mut libc::DIR = unsafe { mem::transmute(fh) };
+    let dir: *mut libc::DIR = unsafe { mem::transmute(fh as usize) };
     if -1 == unsafe { libc::closedir(dir) } {
         Err(io::Error::last_os_error().raw_os_error().unwrap())
     } else {
